requirements.txt
python-osc
infi.systray==0.1.12